LC-4sum [Runtime: N/A; Memory: N/A]

+++ array/4sum.cpp
@@ -23,7 +23,7 @@ public:
                 int right=nums.size()-1;
 
                 while (left<right){
-                    long sum = nums[i]+nums[j]+nums[left]+nums[right];
+                    long long sum = nums[i]+nums[j]+nums[left]+nums[right];
                     if (sum>target){
                         right--;
                     }else if (sum<target){
